update:audio_play_write_blank_raw_ex声明冲突

+++ interface/private_src/luat_audio_ec618.c
@@ -25,7 +25,7 @@ extern void audio_play_tts_default_fun(void *param);
 extern void audio_play_tts_set_resource_ex(void *address, void *sdk_id, void *read_resource_fun);
 extern void audio_play_global_init_ex(audio_play_event_cb_fun_t event_cb, audio_play_data_cb_fun_t data_cb, audio_play_default_fun_t play_file_fun, audio_play_default_fun_t play_tts_fun, void *user_param);
 extern void audio_play_global_init_with_task_priority(audio_play_event_cb_fun_t event_cb, audio_play_data_cb_fun_t data_cb, audio_play_default_fun_t play_file_fun, audio_play_default_fun_t play_tts_fun, void *user_param, uint8_t priority);
-extern int audio_play_write_blank_raw_ex(uint8_t multimedia_id, uint8_t cnt, uint8_t add_font);
+extern int audio_play_write_blank_raw_ex(uint32_t multimedia_id, uint8_t cnt, uint8_t add_font);
 void luat_audio_play_file_default_fun(void *param)
 {
 	audio_play_file_default_fun(param);
